requirements.txt
flask>=2.3.2
Werkzeug>=2.3.6
python-dotenv>=1.0.0

numpy>=1.23.5
opencv-python>=4.7.0.72
six>=1.16.0
tensorflow>=2.12.0
keras>=2.12.0